+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.0.0] - 2021-04-30
 ### Changed
 - Message encoding now defaults to UTF-8
 - Send-HtmlMailMessage now uses Mailozaurr module (which uses MailKit) to send email as Send-MailMessage is being deprecated
@@ -29,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Send-HtmlMailMessage cmdlet
  - Invoke-CommandWithEmailWrapper cmdlet
 
-[Unreleased]: https://github.com/natescherer/PoshEmail/compare/v1.1.0..HEAD
+[Unreleased]: https://github.com/natescherer/PoshEmail/compare/v2.0.0..HEAD
+[2.0.0]: https://github.com/natescherer/PoshEmail/compare/v1.1.0..v2.0.0
 [1.1.0]: https://github.com/natescherer/PoshEmail/compare/v1.0.0..v1.1.0
 [1.0.0]: https://github.com/natescherer/PoshEmail/tree/v1.0.0